✨ add search.routes
