+++ src/routes/index.js
@@ -8,8 +8,6 @@ import features from '~/constantes/features';
 import rotas from '~/constantes/rotas';
 import { FormProvider } from '~/context/FormContext';
 import BemVindo from '~/pages/BemVindo';
-import Buscar from '~/pages/Buscar';
-import BuscarDescription from '~/pages/Buscar/Description';
 import TelaDeCadastro from '~/pages/Cadastro';
 import Descricao from '~/pages/Content/Descricao';
 import Description from '~/pages/Content/Description';
@@ -34,25 +32,9 @@ import CadastroRoutes from './cadastro.routes';
 import FrequenciasStackScreen from './frequencias.routes';
 import LoginStackScreen from './login.routes';
 import PreCadastroRoutes from './preCadastro.routes';
+import SearchStackScreen from './search.routes';
 
 const RootStack = createStackNavigator();
-
-// TODO: Avaliar a remoção deste Feature Toggle
-function EdicaoProfissional(props) {
-  return (
-    <FormProvider>
-      <EdicaoInfoProfissional {...props} />
-    </FormProvider>
-  );
-}
-
-function EdicaoPessoal(props) {
-  return (
-    <FormProvider>
-      <EdicaoInfoPessoal {...props} />
-    </FormProvider>
-  );
-}
 
 const ElmoFunc = props => (
   <FormProvider>
@@ -171,7 +153,7 @@ export default function App({ navigationRef }) {
         />
         <RootStack.Screen
           name={rotas.EDICAO_PROFISSIONAL}
-          component={EdicaoProfissional}
+          component={EdicaoInfoProfissional}
         />
         <RootStack.Screen
           name={rotas.BEM_VINDO}
@@ -180,7 +162,7 @@ export default function App({ navigationRef }) {
         />
         <RootStack.Screen
           name={rotas.EDICAO_INFO_PESSOAIS}
-          component={EdicaoPessoal}
+          component={EdicaoInfoPessoal}
         />
         <RootStack.Screen name={rotas.ELMO} component={ElmoFunc} />
         <RootStack.Screen
@@ -211,24 +193,5 @@ export default function App({ navigationRef }) {
         <RootStack.Screen name="manejoWebview" component={ManejoWebViewPage} />
       </RootStack.Navigator>
     </NavigationContainer>
-  );
-}
-
-const SearchStack = createStackNavigator();
-
-function SearchStackScreen() {
-  return (
-    <SearchStack.Navigator>
-      <SearchStack.Screen
-        name="Buscar"
-        component={Buscar}
-        options={{ headerShown: true }}
-      />
-      <SearchStack.Screen
-        name="Buscar Description"
-        component={BuscarDescription}
-        options={{ headerShown: true }}
-      />
-    </SearchStack.Navigator>
   );
 }
